Split out VerificationReport into a separate mc-attest-verifier-types crate.
