+++ Cargo.toml
@@ -15,6 +15,7 @@ members = [
     "attest/trusted",
     "attest/untrusted",
     "attest/verifier",
+    "attest/verifier/types",
     "common",
     "connection",
     "connection/test-utils",
